Upgrade libuv to joyent/libuv@ce20791

+++ deps/uv/test/benchmark-ping-pongs.c
@@ -124,7 +124,7 @@ static void pinger_shutdown_cb(uv_shutdown_t* req, int status) {
 
 
 static void pinger_read_cb(uv_stream_t* tcp, ssize_t nread, uv_buf_t buf) {
-  unsigned int i;
+  ssize_t i;
   pinger_t* pinger;
 
   pinger = (pinger_t*)tcp->data;
