+++ deps/uv/test/test-ping-pong.c
@@ -94,7 +94,7 @@ static void pinger_write_ping(pinger_t* pinger) {
 
 
 static void pinger_read_cb(uv_stream_t* stream, ssize_t nread, uv_buf_t buf) {
-  unsigned int i;
+  ssize_t i;
   pinger_t* pinger;
 
   pinger = (pinger_t*)stream->data;
